+++ spec/support/test_classes.rb
@@ -99,3 +99,100 @@ class PizzaBasket
   include Basket::Batcher
   basket_options size: 10
 end
+
+class OnAddErrorBasket
+  include Basket::Batcher
+  basket_options size: 2
+
+  def on_add
+    raise "on_add exploded"
+  end
+
+  def on_failure
+  end
+end
+
+class OnSuccessErrorBasket
+  include Basket::Batcher
+  basket_options size: 1
+
+  def perform
+    puts "performing"
+  end
+
+  def on_success
+    raise "on_success exploded"
+  end
+
+  def on_failure
+  end
+end
+
+class MemoryBackendBasket
+  include Basket::Batcher
+  basket_options size: 2, backend: :memory
+
+  def perform
+    puts "memory perform: #{batch}"
+  end
+end
+
+class RedisBackendBasket
+  include Basket::Batcher
+  basket_options size: 2, backend: :redis
+
+  def perform
+    puts "redis perform: #{batch}"
+  end
+end
+
+class DefaultBackendBasket
+  include Basket::Batcher
+  basket_options size: 2
+
+  def perform
+    puts "default perform: #{batch}"
+  end
+end
+
+class MemoryBackendSearchBasket
+  include Basket::Batcher
+  basket_options size: 100, backend: :memory
+
+  def perform
+    puts "memory search perform"
+  end
+end
+
+class RedisBackendSearchBasket
+  include Basket::Batcher
+  basket_options size: 100, backend: :redis
+
+  def perform
+    puts "redis search perform"
+  end
+end
+
+class ConcurrencyTrackingBasket
+  include Basket::Batcher
+  basket_options size: 5
+
+  @perform_count = 0
+  @perform_mutex = Mutex.new
+
+  class << self
+    attr_reader :perform_count, :perform_mutex
+
+    def reset_tracking
+      @perform_mutex.synchronize { @perform_count = 0 }
+    end
+
+    def increment_perform_count
+      @perform_mutex.synchronize { @perform_count += 1 }
+    end
+  end
+
+  def perform
+    self.class.increment_perform_count
+  end
+end